CORE-5406: Rename net.corda package inside CPKs to net.cordapp
Regenerate test resources
