+++ libs/serialization/serialization-amqp/cpk-evolution-newer/src/main/kotlin/net/cordapp/bundle/evolution/newer/SerializableStateToNewerVersion.kt
@@ -1,0 +1,12 @@
+package net.cordapp.bundle.evolution.newer
+
+import net.corda.v5.base.annotations.CordaSerializable
+import java.util.UUID
+
+/**
+ * Serializable class used to test evolution to newer CPKs.
+ */
+@CordaSerializable
+// Uncomment this version of the class when rebuilding test resources
+//data class SerializableStateToNewerVersion(val id: UUID)
+data class SerializableStateToNewerVersion(val id: UUID, val addedField: String?)
