--- libs/serialization/serialization-amqp/cpk-one/src/main/kotlin/net.corda.bundle1/Cash.kt
@@ -1,6 +1,0 @@
-package net.corda.bundle1
-
-import net.corda.v5.base.annotations.CordaSerializable
-
-@CordaSerializable
-data class Cash(val amount: Int)
