+++ libs/serialization/serialization-amqp/cpk-one/src/main/kotlin/net/cordapp/bundle1/Cash.kt
@@ -1,0 +1,6 @@
+package net.cordapp.bundle1
+
+import net.corda.v5.base.annotations.CordaSerializable
+
+@CordaSerializable
+data class Cash(val amount: Int)
