--- libs/serialization/serialization-amqp/cpk-swap-replacement/src/main/kotlin/net/corda/bundle/evolution/different/SerializableStateForDifferentCpk.kt
@@ -1,10 +1,0 @@
-package net.corda.bundle.evolution.different
-
-import net.corda.v5.base.annotations.CordaSerializable
-import java.util.UUID
-
-/**
- * Serializable class used to test that we cannot swap CPKs when deserializing
- */
-@CordaSerializable
-data class SerializableStateForDifferentCpk(val id: UUID)
