--- libs/serialization/serialization-amqp/cpk-three/src/main/java/net/corda/bundle3/package-info.java
@@ -1,4 +1,0 @@
-@Export
-package net.corda.bundle3;
-
-import org.osgi.annotation.bundle.Export;
