+++ libs/serialization/serialization-amqp/cpk-three/src/main/java/net/cordapp/bundle3/package-info.java
@@ -1,0 +1,4 @@
+@Export
+package net.cordapp.bundle3;
+
+import org.osgi.annotation.bundle.Export;
