+++ libs/serialization/serialization-amqp/cpk-three/src/main/java/net/cordapp/bundle5/package-info.java
@@ -1,0 +1,4 @@
+@Export
+package net.cordapp.bundle5;
+
+import org.osgi.annotation.bundle.Export;
