--- libs/serialization/serialization-amqp/cpk-three/src/main/kotlin/net/corda/bundle2/Document.kt
@@ -1,3 +1,0 @@
-package net.corda.bundle2
-
-data class Document constructor(val content: String, val version: Int)
