--- libs/serialization/serialization-amqp/cpk-three/src/main/kotlin/net/corda/bundle5/Container.kt
@@ -1,6 +1,0 @@
-package net.corda.bundle5
-
-import net.corda.v5.base.annotations.CordaSerializable
-
-@CordaSerializable
-class Container<T>(val obj: T)
