+++ libs/serialization/serialization-amqp/cpk-three/src/main/kotlin/net/cordapp/bundle2/Document.kt
@@ -1,0 +1,3 @@
+package net.cordapp.bundle2
+
+data class Document constructor(val content: String, val version: Int)
