+++ libs/serialization/serialization-amqp/cpk-three/src/main/kotlin/net/cordapp/bundle5/Container.kt
@@ -1,0 +1,6 @@
+package net.cordapp.bundle5
+
+import net.corda.v5.base.annotations.CordaSerializable
+
+@CordaSerializable
+class Container<T>(val obj: T)
