--- libs/serialization/serialization-amqp/cpk-two/src/main/kotlin/net/corda/bundle2/Document.kt
@@ -1,6 +1,0 @@
-package net.corda.bundle2
-
-import net.corda.v5.base.annotations.CordaSerializable
-
-@CordaSerializable
-class Document(val content: String)
