+++ libs/serialization/serialization-amqp/cpk-two/src/main/kotlin/net/cordapp/bundle2/Document.kt
@@ -1,0 +1,6 @@
+package net.cordapp.bundle2
+
+import net.corda.v5.base.annotations.CordaSerializable
+
+@CordaSerializable
+class Document(val content: String)
