--- libs/serialization/serialization-amqp/cpk-using-lib/src/main/kotlin/net/corda/bundle/MainBundleItem.kt
@@ -1,12 +1,0 @@
-package net.corda.bundle
-
-import com.example.serialization.PrivateBundleItem
-import net.corda.v5.base.annotations.CordaSerializable
-
-@CordaSerializable
-data class MainBundleItem(val privateBundleItem: PrivateBundleItem) {
-    companion object {
-        @JvmStatic
-        fun newInstance(): MainBundleItem = MainBundleItem(PrivateBundleItem(5))
-    }
-}
