+++ libs/serialization/serialization-amqp/cpk-using-lib/src/main/kotlin/net/cordapp/bundle/MainBundleItem.kt
@@ -1,0 +1,12 @@
+package net.cordapp.bundle
+
+import com.example.serialization.PrivateBundleItem
+import net.corda.v5.base.annotations.CordaSerializable
+
+@CordaSerializable
+data class MainBundleItem(val privateBundleItem: PrivateBundleItem) {
+    companion object {
+        @JvmStatic
+        fun newInstance(): MainBundleItem = MainBundleItem(PrivateBundleItem(5))
+    }
+}
