--- libs/serialization/serialization-kryo/cpks/serializable-cpk-one/src/main/kotlin/net.corda.bundle1/Cash.kt
@@ -1,3 +1,0 @@
-package net.corda.bundle1
-
-data class Cash(val amount: Int)
