+++ libs/serialization/serialization-kryo/cpks/serializable-cpk-one/src/main/kotlin/net/cordapp/bundle1/Cash.kt
@@ -1,0 +1,3 @@
+package net.cordapp.bundle1
+
+data class Cash(val amount: Int)
