--- libs/serialization/serialization-kryo/cpks/serializable-cpk-two/src/main/kotlin/net/corda/bundle2/Document.kt
@@ -1,3 +1,0 @@
-package net.corda.bundle2
-
-class Document(val content: String)
