+++ libs/serialization/serialization-kryo/cpks/serializable-cpk-two/src/main/kotlin/net/cordapp/bundle2/Document.kt
@@ -1,0 +1,3 @@
+package net.cordapp.bundle2
+
+class Document(val content: String)
