+++ processors/db-processor/src/test/kotlin/net/corda/processors/db/internal/reconcile/db/CpiInfoDbReconcilerReaderTest.kt
@@ -48,7 +48,7 @@ class CpiInfoDbReconcilerReaderTest {
             )
         ),
         CordappManifest(
-            "net.corda.Bundle",
+            "net.cordapp.Bundle",
             "1.2.3",
             12,
             34,
